Cleaned up constants and imports in Contacts folder
